Consolidated all files into two files
