+++ storage/remote/azuread/azuread.go
@@ -16,31 +16,50 @@ package azuread
 import (
 	"context"
 	"errors"
+	"fmt"
+	"net/http"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/google/uuid"
 )
 
 const (
-	// Clouds
+	// Clouds.
 	AzureChina      = "AzureChina"
 	AzureGovernment = "AzureGovernment"
 	AzurePublic     = "AzurePublic"
 
-	// Audiences
+	// Audiences.
 	INGESTION_CHINA_AUDIENCE      = "https://monitor.azure.cn//.default"
 	INGESTION_GOVERNMENT_AUDIENCE = "https://monitor.azure.us//.default"
 	INGESTION_PUBLIC_AUDIENCE     = "https://monitor.azure.com//.default"
 )
 
+// AzureADConfig is used to store the config values.
+type AzureADConfig struct {
+	// ClientID is the clientId of the managed identity that is being used to authenticate.
+	ClientID string `yaml:"client_id,omitempty"`
+
+	// Cloud is the Azure cloud in which the service is running. Example: AzurePublic/AzureGovernment/AzureChina.
+	Cloud string `yaml:"cloud,omitempty"`
+}
+
+// azureADRoundTripper is used to store the roundtripper and the tokenprovider.
+type azureADRoundTripper struct {
+	next          http.RoundTripper
+	tokenProvider TokenProvider
+}
+
 // tokenProvider is used to store and retrieve Azure AD accessToken.
 type tokenProvider struct {
 	// token is member used to store the current valid accessToken.
 	token string
-	// mu guards access to token
+	// mu guards access to token.
 	mu  sync.Mutex
 	ctx context.Context
 	// refreshTime is used to store the refresh time of the current valid accessToken.
@@ -53,6 +72,86 @@ type tokenProvider struct {
 // TokenProvider is the interface for Credential client.
 type TokenProvider interface {
 	getAccessToken() (string, error)
+}
+
+// Validate validates config values provided.
+func (c *AzureADConfig) Validate() error {
+	if c.Cloud == "" {
+		return fmt.Errorf("must provide a cloud in the Azure AD config")
+	}
+
+	if c.ClientID == "" {
+		return fmt.Errorf("must provide an Azure Managed Identity client_id in the Azure AD config")
+	}
+
+	_, err := uuid.Parse(c.ClientID)
+
+	if err != nil {
+		return fmt.Errorf("the provided Azure Managed Identity client_id provided is invalid")
+	}
+	return nil
+}
+
+// UnmarshalYAML unmarshal the Azure AD config yaml.
+func (c *AzureADConfig) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	type plain AzureADConfig
+	*c = AzureADConfig{}
+	if err := unmarshal((*plain)(c)); err != nil {
+		return err
+	}
+	return c.Validate()
+}
+
+// NewAzureADRoundTripper creates round tripper adding Azure AD authorization to calls.
+func NewAzureADRoundTripper(cfg *AzureADConfig, next http.RoundTripper) (http.RoundTripper, error) {
+	if next == nil {
+		next = http.DefaultTransport
+	}
+
+	cred, err := newTokenCredential(cfg)
+	if err != nil {
+		return nil, err
+	}
+
+	tokenProvider, err := newTokenProvider(context.Background(), cfg, cred)
+	if err != nil {
+		return nil, err
+	}
+
+	rt := &azureADRoundTripper{
+		next:          next,
+		tokenProvider: tokenProvider,
+	}
+	return rt, nil
+}
+
+// RoundTrip sets Authorization header for requests.
+func (rt *azureADRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
+	accessToken, err := rt.tokenProvider.getAccessToken()
+	if err != nil {
+		return nil, err
+	}
+	bearerAccessToken := "Bearer " + accessToken
+	req.Header.Set("Authorization", bearerAccessToken)
+
+	return rt.next.RoundTrip(req)
+}
+
+// newTokenCredential returns a TokenCredential of different kinds like Azure Managed Identity and Azure AD application.
+func newTokenCredential(cfg *AzureADConfig) (azcore.TokenCredential, error) {
+	cred, err := newManagedIdentityTokenCredential(cfg.ClientID)
+	if err != nil {
+		return nil, err
+	}
+
+	return cred, nil
+}
+
+// newManagedIdentityTokenCredential returns new Managed Identity token credential.
+func newManagedIdentityTokenCredential(managedIdentityClientId string) (azcore.TokenCredential, error) {
+	clientId := azidentity.ClientID(managedIdentityClientId)
+	opts := &azidentity.ManagedIdentityCredentialOptions{ID: clientId}
+	return azidentity.NewManagedIdentityCredential(opts)
 }
 
 // newTokenProvider helps to fetch accessToken for different types of credential. This also takes care of
@@ -91,9 +190,9 @@ func (tokenProvider *tokenProvider) getAccessToken() (string, error) {
 	return tokenProvider.token, nil
 }
 
-// valid checks if the token in the token provider is valid and not expired
+// valid checks if the token in the token provider is valid and not expired.
 func (tokenProvider *tokenProvider) valid() bool {
-	if tokenProvider.token != "" && tokenProvider.refreshTime.After(time.Now().UTC()) {
+	if tokenProvider.refreshTime.After(time.Now().UTC()) {
 		return true
 	} else {
 		return false
@@ -106,6 +205,9 @@ func (tokenProvider *tokenProvider) getToken() error {
 	if err != nil {
 		return err
 	}
+	if len(accessToken.Token) == 0 {
+		return errors.New("Access Token is empty")
+	}
 
 	tokenProvider.token = accessToken.Token
 	err = tokenProvider.updateRefreshTime(accessToken)
@@ -117,9 +219,6 @@ func (tokenProvider *tokenProvider) getToken() error {
 
 // updateRefreshTime handles logic to set refreshTime. The refreshTime is set at half the duration of the actual token expiry.
 func (tokenProvider *tokenProvider) updateRefreshTime(accessToken azcore.AccessToken) error {
-	if len(accessToken.Token) == 0 {
-		return errors.New("Access Token is empty")
-	}
 	tokenExpiryTimestamp := accessToken.ExpiresOn.UTC()
 	deltaExpirytime := time.Now().Add(tokenExpiryTimestamp.Sub(time.Now()) / 2)
 	if deltaExpirytime.After(time.Now().UTC()) {
