--- storage/remote/azuread/credential_mock.go
@@ -1,36 +1,0 @@
-// Copyright 2023 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package azuread
-
-import (
-	"context"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/stretchr/testify/mock"
-)
-
-// Mock azidentity TokenCredential interface
-type mockCredential struct {
-	mock.Mock
-}
-
-func (m *mockCredential) GetToken(ctx context.Context, options policy.TokenRequestOptions) (azcore.AccessToken, error) {
-	args := m.MethodCalled("GetToken", ctx, options)
-	if args.Get(0) == nil {
-		return azcore.AccessToken{}, args.Error(1)
-	}
-
-	return args.Get(0).(azcore.AccessToken), nil
-}
